==> 01_create.ipynb <==
# Create data

The aim of this Capstone project is to consolidate the knowledge, skills and competences acquired by learners during this programme. In this project students will select, specify, design, justify and implement a data analytics investigation that makes use of large data sets. The project will necessarily require data ingestion, wrangling, computation and analyses, and the production and presentation of a project and report. Students are guided by faculty in the selection of a data set, scope of investigation, initial exploratory data analysis, and formulation of hypotheses.

import pandas as pd
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt
import ast

from clean import create_card_df
from clean import create_deck_df

deck_df = create_deck_df()

Clear that unnamed column that snuck in

deck_df.drop(deck_df.columns[deck_df.columns.str.contains('Unnamed',case = False)],axis = 1, inplace = True)

deck_df.head()

print("Total decks:" + str(len(deck_df)))

print("Decks with wins:" + str(len(deck_df.loc[(deck_df['deck_wins'] > 0)])))

print("Decks with wins or losses:" + str(len(deck_df.loc[(deck_df['deck_wins'] > 0) | (deck_df['deck_losses'] > 0)])))

card_df = create_card_df()
print(card_df.iloc[0])

print("Total cards:" + str(len(card_df)))

Remove any blank columns that snuck in.

card_df.drop(card_df.columns[card_df.columns.str.contains('Unnamed',case = False)],axis = 1, inplace = True)

Remove any duplicates so we just have a single entry for every card.

card_df.drop_duplicates(subset=None, keep='first', inplace=True)

card_df.head()

print("Total cards:" + str(len(card_df)))

winning_decks = deck_df.loc[(deck_df['deck_wins'] > 0) | (deck_df['deck_losses'] > 0)]

print("Total decks:" + str(len(winning_decks)))

winning_decks.to_csv("data/winning_decks.csv", header=True, index=False)

card_df.to_csv("data/cards.csv", header=True, index=False)

-- End of File --
