==> 02_tidy.ipynb <==
# Tidy Data

The aim of this Capstone project is to consolidate the knowledge, skills and competences acquired by learners during this programme. In this project students will select, specify, design, justify and implement a data analytics investigation that makes use of large data sets. The project will necessarily require data ingestion, wrangling, computation and analyses, and the production and presentation of a project and report. Students are guided by faculty in the selection of a data set, scope of investigation, initial exploratory data analysis, and formulation of hypotheses.

import pandas as pd
import numpy as np

deck_df = pd.read_csv("data/winning_decks.csv")

card_df = pd.read_csv("data/cards.csv")

There are 6 houses in total: Brobnar, Dis, Sanctum, Mars, Untamed, Shadows, Logos. We want to add those 6 columns and a boolean True of False for whether the deck has that house or not.

deck_df['house_brobnar'] = deck_df['houses'].str.contains('Brobnar')
deck_df['house_dis'] = deck_df['houses'].str.contains('Dis')
deck_df['house_sanctum'] = deck_df['houses'].str.contains('Sanctum')
deck_df['house_mars'] = deck_df['houses'].str.contains('Mars')
deck_df['house_untamed'] = deck_df['houses'].str.contains('Untamed')
deck_df['house_shadows'] = deck_df['houses'].str.contains('Shadows')
deck_df['house_logos'] = deck_df['houses'].str.contains('Logos')

deck_df.head()

deck_df = deck_df.drop("houses", axis=1)

deck_df.head()

Next we need to look at the cards.

card_df.head()

For some cards they only appear once.

card_df.loc[card_df['card_title']=='Ixxyxli Fixfinger']

Other cards however appear under multiple card_ids, this is because they have special versions that appear in other houses. So if we look at the card 'Bait and Switch' we can see it appears a total of 7 times, one for each house.

card_df.loc[card_df['card_title']=='Bait and Switch']

We a column for each of the unique cards name which are a total of:

len(card_df.groupby('card_title')['card_title'].nunique())

So we need to add 574 columns. Lets start by getting the list of names.

series_of_cards = card_df['card_title']
list_of_cards = list(dict.fromkeys(series_of_cards))
list_of_cards.sort()
list_of_cards

We would like to have a count of the the number of each card in the deck. This is problematic as there are multiple card_ids for each card name. That would be tricky to match. Instead what we'll try do, is add a new column which has the card names instead of the card ids.

card_df.loc[(card_df['card_id'] == deck_df.iloc[0]['deck_list'][2:38])]['card_title'].item()

def change_card_ids_to_names(card_list):
    new_card_list = []
    start = 2
    end = 38
    for i in range(36):
        card = card_df.loc[(card_df['card_id'] == card_list[start:end])]['card_title'].item()
        new_card_list.append(card)
        start = start + 40
        end = end + 40
    return(new_card_list)

deck_df.iloc[0]['deck_list']

card_df.iloc[0]['card_id']

change_card_ids_to_names(deck_df.iloc[1]['deck_list'])

This next bit takes some time. The list of card_ids is being replaced with a list of card names. This will make it easier to work with going forward.

%%time
deck_df['card_name_list'] = deck_df['deck_list'].apply(change_card_ids_to_names)

deck_df.iloc[0]

The card name [REDACTED] causes problems later, so we rename to avoid this.

list_of_cards[568] = "REDACTED"

Now we can use a similar method as we did with the houses using a check to see if the name is included in the field. This cycles through the list of 574 card columns and counts the number of times the card name appears in the total card list string.

for n in list_of_cards:
    deck_df[n] = deck_df['card_name_list'].astype(str).str.count(n)

deck_df.iloc[0]

Next we want to assign a 'score' to each deck, this will be our target variable. In this case we are going to take the number of wins and subtract the number of losses.

deck_df.describe()

deck_df['score'] = deck_df['deck_wins'] - deck_df['deck_losses']

deck_df.iloc[0]['score']

Almost there. Now we're going to tidy up a little. We remove the columns we no longer need and we're adjusting the name on the expansion which currently is a numeric.

deck_df.loc[deck_df.deck_expansion == 341, 'expansion_coa'] = True #'Call of the Archons'
deck_df.loc[deck_df.deck_expansion != 341, 'expansion_coa'] = False
deck_df.loc[deck_df.deck_expansion == 435, 'expansion_aoa'] = True #'Age of Ascension'
deck_df.loc[deck_df.deck_expansion != 435, 'expansion_aoa'] = False

deck_df = deck_df.drop("deck_list", axis=1)
deck_df = deck_df.drop("card_name_list", axis=1)

deck_df.loc[0]

deck_df.to_csv("data/data_decks_cleaned.csv", header=True, index=False)

Data prep ends here

------------------------------------------------------
