==> 04_recommender.ipynb <==
# Recommender

With the model created, we want to be able to take the url of two decks and see which one is better

import pandas as pd
import numpy as np
from sklearn import linear_model
import pickle
from load import load_single_deck
from load import decode_single_deck

Using pickle we can load the model

filename = 'finalized_linear_model.sav'
lm = pickle.load(open(filename, 'rb'))

These are the two decks we're going to look at, these are in the format the use would get from the website.

deck1 = "https://www.keyforgegame.com/deck-details/f62c8aec-ab9f-4623-947d-8cfa412a0290"
deck2 = "https://www.keyforgegame.com/deck-details/8af0f140-b4a6-4490-a367-2cc1aa0d6b10"

Next we need to convert this into a format that the model can use.

deck1_code = deck1[-36:]
deck2_code = deck2[-36:]

deck1_data = load_single_deck(deck1_code)
deck2_data = load_single_deck(deck2_code)

deck1_df = decode_single_deck(deck1_data)
deck2_df = decode_single_deck(deck2_data)
deck_df = pd.concat([deck1_df, deck2_df], axis=0)
deck_df

card_df = pd.read_csv("data/cards.csv")

if 'Brobnar' in deck_df['houses'].values:
    deck_df['house_brobnar'] = True
else:
    deck_df['house_brobnar'] = False
if 'Dis' in deck_df['houses'].values:
    deck_df['house_dis'] = True
else:
    deck_df['house_dis'] = False
if 'Sanctum' in deck_df['houses'].values:
    deck_df['house_sanctum'] = True
else:
    deck_df['house_sanctum'] = False
if 'Mars' in deck_df['houses'].values:
    deck_df['house_mars'] = True
else:
    deck_df['house_mars'] = False
if 'Untamed' in deck_df['houses'].values:
    deck_df['house_untamed'] = True
else:
    deck_df['house_untamed'] = False
if 'Shadows' in deck_df['houses'].values:
    deck_df['house_shadows'] = True
else:
    deck_df['house_shadows'] = False
if 'Logos' in deck_df['houses'].values:
    deck_df['house_logos'] = True
else:
    deck_df['house_logos'] = False

deck_df = deck_df.drop("houses", axis=1)
deck_df.drop(deck_df.columns[deck_df.columns.str.contains('Unnamed',case = False)],axis = 1, inplace = True)

deck_df.head()

def change_card_array_ids_to_names(card_list):
    new_card_list = []
    for i in card_list:
        card = card_df.loc[(card_df['card_id'] == i)]['card_title'].item()
        new_card_list.append(card)
    return(new_card_list)

series_of_cards = card_df['card_title']
list_of_cards = list(dict.fromkeys(series_of_cards))
list_of_cards.sort()
list_of_cards[568] = "REDACTED"
list_of_cards

deck_df['card_name_list'] = deck_df['deck_list'].apply(change_card_array_ids_to_names)

deck_df.head()

for n in list_of_cards:
    deck_df[n] = deck_df['card_name_list'].astype(str).str.count(n)

deck_df.loc[deck_df.deck_expansion == 435, 'expansion_coa'] = True #'Call of the Archons'
deck_df.loc[deck_df.deck_expansion == 341, 'expansion_coa'] = False
deck_df.loc[deck_df.deck_expansion == 341, 'expansion_aoa'] = True #'Age of Asecension'
deck_df.loc[deck_df.deck_expansion == 435, 'expansion_aoa'] = False

deck_df.head()

deck_df = deck_df.drop("deck_list", axis=1)
deck_df = deck_df.drop("card_name_list", axis=1)

X = deck_df
X = X.drop("deck_id", axis=1)
X = X.drop("deck_name", axis=1)
X = X.drop("deck_wins", axis=1)
X = X.drop("deck_losses", axis=1)
X = X.drop("house_shadows", axis=1)
X = X.drop("house_mars", axis=1)
X = X.drop("house_untamed", axis=1)
X = X.drop("house_logos", axis=1)
X = X.drop("house_dis", axis=1)
X = X.drop("house_sanctum", axis=1)
X = X.drop("house_brobnar", axis=1)
X = X.drop("deck_expansion", axis=1)
X = X.drop("expansion_coa", axis=1)
X = X.drop("expansion_aoa", axis=1)

X

lm

predictions = lm.predict(X)
print(predictions)

deck_df['deck_name']
